manage/requirements.txt
Django>=5.1.6
djangorestframework>=3.13
psycopg2-binary>=2.9.10
drf-yasg>=1.21.10
djangorestframework-simplejwt>=5.0